SJ-2107 Fixed variable types

+++ core/sj-crud-rest/src/main/scala/com/bwsw/sj/crud/rest/entities/module/InputInstanceMetadata.scala
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
   * @author Kseniya Tomskikh
   */
 class InputInstanceMetadata extends InstanceMetadata {
-  @JsonProperty("lookup-history") var lookupHistory: Long = 0
-  @JsonProperty("queue-max-size") var queueMaxSize: Long = 0
+  @JsonProperty("lookup-history") var lookupHistory: Int = 0
+  @JsonProperty("queue-max-size") var queueMaxSize: Int = 0
   @JsonProperty("eviction-policy") var evictionPolicy: String = null
 }
